Add support for building against a Yocto SDK
In a Yocto SDK environment the following relevant environment variables
are set:

 - SDKTARGETSYSROOT: points to the target sysroot.
 - CLANGCC/CLANGCXX: the command line to use for cross-compiling with
   clang to the target.
 - CC/CXX: the gcc command line for cross-compiling.

This patch introduces a dependency to these two variables, as preferred
sysroot and clang choices after the macOS specific SDKROOT and CC/CXX.

Another aspect to the Yocto build is that the target name diverges from
cargo.  For example what's aarch64-unknown-linux-gnu with cargo might be
aarch64-poky-linux (vendor set, no abi specified).

On a clang level this is dealt with by CLANGCC/CLANGCXX including
--target=aarch64-poky-linux for example.

Unfortunately the Skia build also adds --target to the command line,
based on Cargo's target. In the final command line when compiling Skia,
compiling bindings.cpp or generating the bindings, clang's attempt at
locating the correct gcc installation for libstdc++ headers fails,
because --target=aarch64-unknown-linux-gnu does not match the
compiled-in paths that clang would look for if the target was
aarch64-poky-linux (as per original command line).

In order to fix that, this patch also extracts any existing
--target=$foo from the build configuration and uses it throughout in
favor of cargo::target() if provided.