Fix compilation with meta-rust in Yocto
When compiling in Yocto with the upstream meta-rust instead of meta-rust-bin,
the Yocto target specs are used for compilation. That means the target strings
aren't necessarily triplets (or longer) in the form of x86_64-unknown-linux-gnu, but they
can be also short like x86_64-linux.

Therefore adjust parse_target() to accomodate that format.