Fix memory corruption when creating YUVABackendTextures
The C++ constructor expects an array of GrBackendTexture. Since #751 the
BackendTexture type in Rust is not a GrBackendTexture anymore though,
but a refhandle, so we were passing a pointer.

Work around this by not calling the C++ constructor directly anymore but
use a helper function in gpu.cpp that convers the array of pointers to
an array of GrBackendTexture objects.