Fix build on Windows from within vcvars shell
vcvars on my machine (and in the GitHub actions VMs) sets VCINSTALLDIR
to something like C:\Program Files\Microsoft Visual Studio\2022\Community\VC\,
where the trailing backslash conflicts with the quote that is appended.

That results in the following GN args

 ... win_vc="C:\Program Files\Microsoft Visual Studio\2022\Community\VC\" clang_win

and this way win_vc is not properly terminated.