Add prebuilt skia binary of Metal+OpenGL on macOS.
Currently we only have Metal only or OpenGL only but not both
together on macOS. However, there are all three combinations with
D3D on Windows. For the same reason, sometimes developers want to
build a program using both OpenGL and Metal. So we add this
combination of features to make it easier.